refactor(wiki): adapt for crawlR wiki
